Time: 1 ms (61.11%), Space: 65.1 MB (66.26%) - LeetHub

+++ 0713-subarray-product-less-than-k/0713-subarray-product-less-than-k.cpp
@@ -1,14 +1,15 @@
 class Solution {
 public:
     int numSubarrayProductLessThanK(vector<int>& nums, int k) {
+        // uses expand and shrink window approach here : 
         int start = 0;
         int end = 0;
+
         long long product = 1;
-        int count = 0;
+        long long count = 0;
 
         while(end < nums.size()) {
             product *= nums[end];
-
             while(start <= end && product >= k) {
                 product /= nums[start];
                 start++;
@@ -16,8 +17,9 @@ public:
 
             count += end-start+1;
             end++;
-
         }
+
         return count;
+
     }
 };
